# apps/server/docs/swagger.yaml
basePath: /api/v1
definitions:
  catalog.CodeName:
    properties:
      code:
        example: FRIDGE
        type: string
      name:
        example: 냉장
        type: string
    type: object
  catalog.FoodCategoryCode:
    enum:
    - DAIRY
    - MEAT
    - SEAFOOD
    - VEGETABLE
    - FRUIT
    - KIMCHI_SIDE
    - FROZEN_FOOD
    - SAUCE
    - DRINK
    - ETC
    type: string
    x-enum-varnames:
    - FoodCategoryDairy
    - FoodCategoryMeat
    - FoodCategorySeafood
    - FoodCategoryVegetable
    - FoodCategoryFruit
    - FoodCategoryKimchiSide
    - FoodCategoryFrozenFood
    - FoodCategorySauce
    - FoodCategoryDrink
    - FoodCategoryEtc
...
  handlers.CreatePantryItemRequest:
    properties:
      expiresAt:
        example: "2026-02-21"
        type: string
      foodCategoryCode:
        allOf:
        - $ref: '#/definitions/catalog.FoodCategoryCode'
        description: '식품 카테고리 코드: DAIRY, MEAT, SEAFOOD, VEGETABLE, FRUIT, KIMCHI_SIDE,
          FROZEN_FOOD, SAUCE, DRINK, ETC'
        enum:
        - DAIRY
        - MEAT
        - SEAFOOD
        - VEGETABLE
        - FRUIT
        - KIMCHI_SIDE
        - FROZEN_FOOD
        - SAUCE
        - DRINK
        - ETC
        example: DAIRY
      foodId:
        example: 00000000-0000-0000-0000-000000000002
        type: string
      inputName:
        example: 우유
        type: string
      purchasedAt:
        example: "2026-02-14"
        type: string
      storageSpaceId:
        example: 00000000-0000-0000-0000-000000000005
        type: string
    type: object
...
  handlers.ErrorResponse:
    properties:
      message:
        example: not implemented
        type: string
    type: object
...
  handlers.UpdatePantryItemRequest:
    properties:
      expiresAt:
        example: "2026-02-22"
        type: string
      inputName:
        example: 딸기 우유
        type: string
      notes:
        example: 아침 시리얼용
        type: string
      purchasedAt:
        example: "2026-02-15"
        type: string
      storageSpaceId:
        example: 00000000-0000-0000-0000-000000000005
        type: string
    type: object
...
info:
  contact: {}
  description: API for naengrijeongdon inventory management.
  title: NaengriJeongdon API
  version: 0.1.0
paths:
...
  /foods/{foodId}:
    get:
      parameters:
      - description: 식품 ID
        in: path
        name: foodId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - AccessTokenAuth: []
      summary: 식품 상세 조회
      tags:
      - Foods
  /foods/{foodId}/predict-expiry:
    get:
      parameters:
      - description: 식품 ID
        in: path
        name: foodId
        required: true
        type: string
      - description: 보관 공간 ID
        example: 00000000-0000-0000-0000-000000000005
        in: query
        name: storageSpaceId
        required: true
        type: string
      - description: 구매일(YYYY-MM-DD)
        example: "2026-02-18"
        in: query
        name: purchasedAt
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - AccessTokenAuth: []
      summary: 유통기한 예측 미리보기
      tags:
      - Foods
  /foods/autocomplete:
    get:
      parameters:
      - description: 검색어
        example: 우유
        in: query
        name: q
        required: true
        type: string
      - default: 10
        description: 최대 개수
        in: query
        maximum: 50
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - AccessTokenAuth: []
      summary: 식품 자동완성 조회
      tags:
      - Foods
...
  /metadata/food-categories:
    get:
      description: '코드 설명: DAIRY(유제품), MEAT(육류), SEAFOOD(해산물), VEGETABLE(채소), FRUIT(과일),
        KIMCHI_SIDE(김치/반찬), FROZEN_FOOD(냉동식품), SAUCE(소스/조미료), DRINK(음료), ETC(기타)'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/catalog.CodeName'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - AccessTokenAuth: []
      summary: 식품 카테고리 목록 조회
      tags:
      - Metadata
  /metadata/storage-types:
    get:
      description: '코드 설명: FRIDGE(냉장), FREEZER(냉동), KIMCHI(김치냉장), ROOM(실온)'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/catalog.CodeName'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - AccessTokenAuth: []
      summary: 보관 타입 목록 조회
      tags:
      - Metadata
  /pantry-items:
    get:
      description: 'status 코드: ALL(전체), EXPIRED(만료), TODAY(오늘 마감), DUE_3_DAYS(3일 이내),
        DUE_7_DAYS(7일 이내), SAFE(여유 있음), NO_EXPIRY(마감일 없음). sort 코드: EXPIRES_ASC(마감일
        빠른순), CREATED_DESC(등록 최신순).'
      parameters:
      - default: ALL
        description: 상태 필터
        enum:
        - ALL
        - EXPIRED
        - TODAY
        - DUE_3_DAYS
        - DUE_7_DAYS
        - SAFE
        - NO_EXPIRY
        in: query
        name: status
        type: string
      - description: 공간 ID
        example: 00000000-0000-0000-0000-000000000005
        in: query
        name: storageSpaceId
        type: string
      - description: 검색어
        example: 우유
        in: query
        name: q
        type: string
      - default: EXPIRES_ASC
        description: 정렬
        enum:
        - EXPIRES_ASC
        - CREATED_DESC
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - AccessTokenAuth: []
      summary: 내 식재료 목록 조회
      tags:
      - PantryItems
    post:
      consumes:
      - application/json
      description: 'foodCategoryCode 코드: DAIRY(유제품), MEAT(육류), SEAFOOD(해산물), VEGETABLE(채소),
        FRUIT(과일), KIMCHI_SIDE(김치/반찬), FROZEN_FOOD(냉동식품), SAUCE(소스/조미료), DRINK(음료),
        ETC(기타). 실시간 목록은 GET /api/v1/metadata/food-categories 참고.'
      parameters:
      - description: 등록 요청
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreatePantryItemRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - AccessTokenAuth: []
      summary: 식재료 등록
      tags:
      - PantryItems
  /pantry-items/{itemId}:
    delete:
      parameters:
      - description: 식재료 ID
        in: path
        name: itemId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - AccessTokenAuth: []
      summary: 식재료 삭제
      tags:
      - PantryItems
    get:
      parameters:
      - description: 식재료 ID
        in: path
        name: itemId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - AccessTokenAuth: []
      summary: 식재료 단건 조회
      tags:
      - PantryItems
    patch:
      consumes:
      - application/json
      parameters:
      - description: 식재료 ID
        in: path
        name: itemId
        required: true
        type: string
      - description: 수정 요청
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdatePantryItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - AccessTokenAuth: []
      summary: 식재료 수정
      tags:
      - PantryItems
  /pantry-items/{itemId}/ack-storage-warning:
    post:
      parameters:
      - description: 식재료 ID
        in: path
        name: itemId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - AccessTokenAuth: []
      summary: 보관 방식 경고 확인 처리
      tags:
      - PantryItems
...
securityDefinitions:
  AccessTokenAuth:
    description: 'Bearer access token. Example: Bearer eyJ...'
    in: header
    name: Authorization
    type: apiKey
  RefreshTokenAuth:
    description: 'Bearer refresh token. Example: Bearer eyJ...'
    in: header
    name: X-Refresh-Token
    type: apiKey
swagger: "2.0"